# src/main/resources/META-INF/openapi.json
{
    "swagger": "2.0",
    "info": {
        "title": "Payment Request (PSD2, STET)",
        "version": "1.3.0",
        "description": "Merchant Initiates a payment request to customer account, so that the customer can give their consent to complete the transaction at merchant web site.",
        "contact": {
            "name": "API Support",
            "url": "https://community.fusionfabric.cloud/index.html"
        },
        "x-finastra-category": "Consumer Banking",
        "x-finastra-subcategory": "Payments",
        "x-finastra-short-description": "Merchant Initiates a payment request to customer account, so that the customer can give their consent to complete the transaction at merchant web site.",
        "x-finastra-tags": [
            "Fusion Digital Channels",
            "Payment",
            "Excluding North America"
        ],
        "x-finastra-channel-type": "B2B",
        "x-finastra-maturity-level": "GA",
        "x-finastra-audience": "PUBLIC",
        "x-finastra-implementability": "INTERNAL",
        "x-finastra-icon": "Icons_9m.svg",
        "x-finastra-data-classification": [
            "Restricted"
        ],
        "x-finastra-commitId": "07970c287ef26f9cfec813e47ae123470151c343"
    },
    "host": "api.fusionfabric.cloud",
    "basePath": "/retail/psd2/stet/pisp/v1.3",
    "schemes": [
        "https"
    ],
    "paths": {
        "/payment-requests": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json;charset=utf-8",
                    "application/hal+json;charset=utf-8"
                ],
                "tags": [
                    "PISP"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "description": "Access token to be passed as a header",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "x-finastra-data-classification": [
                            "Restricted"
                        ]
                    },
                    {
                        "name": "paymentRequest",
                        "description": "ISO20022 based payment Initiation Request",
                        "schema": {
                            "$ref": "#/definitions/PaymentRequestResource"
                        },
                        "in": "body",
                        "required": true
                    },
                    {
                        "name": "Signature",
                        "description": "http-signature of the request (cf. https://datatracker.ietf.org/doc/draft-cavage-http-signatures/) The keyId part of the header should be formatted as follows   keiId=\"SN=XXX,CA=YYYYYYYYYYYYYYYY\" where   \"XXX\" is the serial number, in hexadecimal without any prefix (like 0x), of the QSEAL certificate whose relevant private key was used for signing   \"YYYYYYYYYYYYYYYY\" is the the Issuer DN, full Distinguished Name of the Certification Authority having issued this certificate  HTTP400 will be returned by the server in case of invalid or absent signature",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "x-finastra-data-classification": [
                            "Restricted"
                        ]
                    },
                    {
                        "name": "X-Request-ID",
                        "description": "Correlation header to be set in a request and retrieved in the relevant response",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "PSU-IP-Address",
                        "description": "IP address used by the PSU's terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string",
                        "x-finastra-data-classification": [
                            "Restricted"
                        ]
                    },
                    {
                        "name": "PSU-IP-Port",
                        "description": "IP port used by the PSU's terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-HTTP-Method",
                        "description": "Http method for the most relevant PSU's terminal request to the TTP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Date",
                        "description": "Timestamp of the most relevant PSU's terminal request to the TTP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-GEO-Location",
                        "description": "Geographical location of the PSU as provided by the PSU mobile terminal if any to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-User-Agent",
                        "description": "\"User-Agent\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Referer",
                        "description": "\"Referer\" header field sent by the PSU terminal when connecting to the TPP. Notice that an initial typo in RFC 1945 specifies that \"referer\" (incorrect spelling) is to be used. The correct spelling \"referrer\" can be used but might not be understood.",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept",
                        "description": "\"Accept\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept-Charset",
                        "description": "\"Accept-Charset\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept-Encoding",
                        "description": "\"Accept-Encoding\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept-Language",
                        "description": "\"Accept-Language\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Digest",
                        "description": "Digest of the body",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/HalPaymentTransferRequestCreation"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/HalPaymentTransferRequestCreation"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized, authentication failure.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden, authentication successful but access to resource is not allowed.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "408": {
                        "description": "Request Timeout.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "429": {
                        "description": "Too many requests.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "503": {
                        "description": "Service unavailable.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                },
                "operationId": "paymentRequestsPost_1",
                "summary": "Initiate payment request from the customer bank account on behalf of a merchant",
                "description": "The third-party provider (PISP) send a payment request to the customer's bank in a standard structure (ISO:20022 pain.013.014) to complete the commercial transaction.",
                "x-finastra-data-classification": [
                    "Restricted"
                ]
            }
        },
        "/payment-requests/{paymentRequestResourceId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json;charset=utf-8",
                    "application/hal+json;charset=utf-8"
                ],
                "tags": [
                    "PISP"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "description": "Access token to be passed as a header",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "x-finastra-data-classification": [
                            "Restricted"
                        ]
                    },
                    {
                        "name": "paymentRequestResourceId",
                        "description": "Identification of the Payment Request Resource",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Signature",
                        "description": "http-signature of the request (cf. https://datatracker.ietf.org/doc/draft-cavage-http-signatures/) The keyId part of the header should be formatted as follows   keiId=\"SN=XXX,CA=YYYYYYYYYYYYYYYY\" where   \"XXX\" is the serial number, in hexadecimal without any prefix (like 0x), of the QSEAL certificate whose relevant private key was used for signing   \"YYYYYYYYYYYYYYYY\" is the the Issuer DN, full Distinguished Name of the Certification Authority having issued this certificate  HTTP400 will be returned by the server in case of invalid or absent signature",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "x-finastra-data-classification": [
                            "Restricted"
                        ]
                    },
                    {
                        "name": "X-Request-ID",
                        "description": "Correlation header to be set in a request and retrieved in the relevant response",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "PSU-IP-Address",
                        "description": "IP address used by the PSU's terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string",
                        "x-finastra-data-classification": [
                            "Restricted"
                        ]
                    },
                    {
                        "name": "PSU-IP-Port",
                        "description": "IP port used by the PSU's terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-HTTP-Method",
                        "description": "Http method for the most relevant PSU's terminal request to the TTP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Date",
                        "description": "Timestamp of the most relevant PSU's terminal request to the TTP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-GEO-Location",
                        "description": "Geographical location of the PSU as provided by the PSU mobile terminal if any to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-User-Agent",
                        "description": "\"User-Agent\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Referer",
                        "description": "\"Referer\" header field sent by the PSU terminal when connecting to the TPP. Notice that an initial typo in RFC 1945 specifies that \"referer\" (incorrect spelling) is to be used. The correct spelling \"referrer\" can be used but might not be understood.",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept",
                        "description": "\"Accept\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept-Charset",
                        "description": "\"Accept-Charset\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept-Encoding",
                        "description": "\"Accept-Encoding\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "PSU-Accept-Language",
                        "description": "\"Accept-Language\" header field sent by the PSU terminal when connecting to the TPP",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Digest",
                        "description": "Digest of the body",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieval of the previously posted Payment Request",
                        "schema": {
                            "$ref": "#/definitions/HalPaymentRequest"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized, authentication failure.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden, authentication successful but access to resource is not allowed.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Not found, no request available.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "408": {
                        "description": "Request Timeout.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "429": {
                        "description": "Too many requests.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "503": {
                        "description": "Service unavailable.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                },
                "operationId": "paymentRequestsGet",
                "summary": "Retrieve payment request status and details",
                "description": "A third-party provider (PISP) request payment status from the bank (ASPSP - Account Servicing Payment Service Provider)",
                "x-finastra-data-classification": [
                    "Restricted"
                ]
            }
        }
    },
    "definitions": {
...
        "CreditTransferTransaction": {
            "title": "CreditTransferTransaction",
            "description": "ISO20022: Payment processes required to transfer cash from the debtor to the creditor.",
            "required": [
                "instructedAmount",
                "paymentId",
                "remittanceInformation",
                "requestedExecutionDate"
            ],
            "type": "object",
            "properties": {
                "instructedAmount": {
                    "$ref": "#/definitions/AmountType"
                },
                "paymentId": {
                    "$ref": "#/definitions/PaymentIdentification"
                },
                "regulatoryReportingCode": {
                    "description": "Information needed due to regulatory and statutory requirements.  Economical codes to be used are provided by the National Competent Authority ",
                    "type": "string"
                },
                "remittanceInformation": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "x-finastra-data-classification": [
                        "Restricted"
                    ]
                },
                "requestedExecutionDate": {
                    "format": "date-time",
                    "description": "ISO20022: Date at which the initiating party requests the clearing agent to process the payment.  ",
                    "type": "string"
                },
                "statusReasonInformation": {
                    "enum": [
                        "AC01",
                        "AC04",
                        "AC06",
                        "AG01",
                        "CUST",
                        "FF01",
                        "FRAD",
                        "MS03",
                        "NOAS",
                        "RR01",
                        "RR03",
                        "RR04",
                        "RR12"
                    ],
                    "type": "string"
                },
                "transactionStatus": {
                    "enum": [
                        "RJCT",
                        "PDNG",
                        "ACSP",
                        "ACSC"
                    ],
                    "type": "string"
                }
            }
        },
        "ErrorModel": {
            "title": "ErrorModel",
            "description": "Generic error report structure",
            "required": [
                "message",
                "status"
            ],
            "type": "object",
            "properties": {
                "error": {
                    "description": "HTTP error text",
                    "type": "string"
                },
                "message": {
                    "description": "HTTP textual reason phrase",
                    "type": "string"
                },
                "path": {
                    "description": "Relevant path that was used",
                    "type": "string"
                },
                "status": {
                    "format": "int32",
                    "description": "HTTP error code",
                    "type": "integer"
                },
                "timestamp": {
                    "format": "date-time",
                    "description": "current timestamp",
                    "type": "string"
                }
            }
        },
...
        "HalPaymentRequest": {
            "title": "HalPaymentRequest",
            "description": "HYPERMEDIA structure used for returning the original Payment Request to the PISP",
            "required": [
                "_links",
                "paymentRequest"
            ],
            "type": "object",
            "properties": {
                "_links": {
                    "$ref": "#/definitions/PaymentTransferRequestLinks"
                },
                "paymentRequest": {
                    "$ref": "#/definitions/PaymentRequestResource"
                }
            }
        },
        "HalPaymentTransferRequestCreation": {
            "title": "HalPaymentTransferRequestCreation",
            "description": "data forwarded by the ASPSP top the PISP after creation of the Payment or Transfer Request resource creation ",
            "type": "object",
            "properties": {
                "_links": {
                    "$ref": "#/definitions/PaymentTransferRequestResourceCreationLinks"
                },
                "appliedScaApproach": {
                    "enum": [
                        "REDIRECT",
                        "DECOUPLED",
                        "EMBEDDED"
                    ],
                    "type": "string"
                }
            }
        },
        "PartyIdentification": {
            "title": "PartyIdentification",
            "description": "API : Description of a Party which can be either a person or an organization.",
            "required": [
                "name"
            ],
            "type": "object",
            "properties": {
                "name": {
                    "description": "ISO20022: Name by which a party is known and which is usually used to identify that party. ",
                    "type": "string",
                    "x-finastra-data-classification": [
                        "Restricted"
                    ]
                },
                "organisationId": {
                    "$ref": "#/definitions/GenericIdentification",
                    "description": "Unique and unambiguous way to identify an organisation."
                },
                "postalAddress": {
                    "$ref": "#/definitions/PostalAddress"
                },
                "privateId": {
                    "$ref": "#/definitions/GenericIdentification",
                    "description": "Unique and unambiguous identification of a person."
                }
            }
        },
...
        "PaymentRequestResource": {
            "title": "PaymentRequestResource",
            "description": "ISO20022: The PaymentRequestResource message is sent by the Creditor sending party to the Debtor receiving party, directly or through agents. It is used by a Creditor to request movement of funds from the debtor account to a creditor.",
            "required": [
                "creationDateTime",
                "creditTransferTransaction",
                "creditor",
                "creditorAccount",
                "initiatingParty",
                "numberOfTransactions",
                "paymentInformationId",
                "paymentTypeInformation",
                "supplementaryData"
            ],
            "type": "object",
            "properties": {
                "chargeBearer": {
                    "description": "- ISO2002: Specifies which party/parties will bear the charges associated with the processing of the payment transaction. - API: Set to \"SLEV\" or omitted",
                    "enum": [
                        "SLEV"
                    ],
                    "type": "string"
                },
                "creationDateTime": {
                    "format": "date-time",
                    "description": "ISO20022: Date and time at which a (group of) payment instruction(s) was created by the instructing party.",
                    "type": "string"
                },
                "creditTransferTransaction": {
                    "description": "ISO20022: Payment processes required to transfer cash from the debtor to the creditor.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CreditTransferTransaction"
                    },
                    "x-finastra-data-classification": [
                        "Restricted"
                    ]
                },
                "creditor": {
                    "$ref": "#/definitions/PartyIdentification"
                },
                "creditorAccount": {
                    "$ref": "#/definitions/AccountIdentification"
                },
                "creditorAgent": {
                    "$ref": "#/definitions/FinancialInstitutionIdentification"
                },
                "debtor": {
                    "$ref": "#/definitions/PartyIdentification"
                },
                "debtorAccount": {
                    "$ref": "#/definitions/AccountIdentification"
                },
                "debtorAgent": {
                    "$ref": "#/definitions/FinancialInstitutionIdentification"
                },
                "initiatingParty": {
                    "$ref": "#/definitions/PartyIdentification"
                },
                "numberOfTransactions": {
                    "format": "int32",
                    "description": "ISO20022: Number of individual transactions contained in the message. ",
                    "type": "integer"
                },
                "paymentInformationId": {
                    "description": "ISO20022 : Reference assigned by a sending party to unambiguously identify the payment information block within the message. ",
                    "type": "string"
                },
                "paymentInformationStatus": {
                    "enum": [
                        "ACCP",
                        "ACSC",
                        "ACSP",
                        "ACTC",
                        "ACWC",
                        "ACWP",
                        "RCVD",
                        "PNDG",
                        "RJCT"
                    ],
                    "type": "string"
                },
                "paymentTypeInformation": {
                    "$ref": "#/definitions/PaymentTypeInformation"
                },
                "purpose": {
                    "enum": [
                        "ACCT",
                        "CASH",
                        "COMC",
                        "CPKC",
                        "TRPT"
                    ],
                    "type": "string"
                },
                "resourceId": {
                    "description": "API: Identifier assigned by the ASPSP for further use of the created resource through API calls ",
                    "type": "string"
                },
                "statusReasonInformation": {
                    "enum": [
                        "AC01",
                        "AC04",
                        "AC06",
                        "AG01",
                        "CUST",
                        "FF01",
                        "FRAD",
                        "MS03",
                        "NOAS",
                        "RR01",
                        "RR03",
                        "RR04",
                        "RR12"
                    ],
                    "type": "string"
                },
                "supplementaryData": {
                    "$ref": "#/definitions/SupplementaryData"
                },
                "ultimateCreditor": {
                    "$ref": "#/definitions/PartyIdentification"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Oauth2": {
            "flow": "application",
            "tokenUrl": "https://api.fusionfabric.cloud/login/v1/sandbox/oidc/token",
            "scopes": {

            },
            "type": "oauth2",
            "description": "Oauth 2.0 authentication"
        }
    },
    "security": [
        {
            "Oauth2": [
            ]
        }
    ],
...
}